Deleted redundant class and migrated methods to WCStatsSqlUtils
